+++ src/components/ScaleRule.vue
@@ -3,9 +3,13 @@ import type { Scale } from '@/scale'
 import { computed } from 'vue'
 import { mmod, valueToCents } from 'xen-dev-utils'
 
-const props = defineProps<{
-  scale: Scale
-}>()
+const props = withDefaults(
+  defineProps<{
+    scale: Scale
+    orientation: 'horizontal' | 'vertical'
+  }>(),
+  { orientation: 'horizontal' }
+)
 
 const ticksAndColors = computed(() => {
   const equaveCents = valueToCents(props.scale.equaveRatio)
@@ -32,7 +36,7 @@ const ticksAndColors = computed(() => {
 </script>
 
 <template>
-  <svg width="100%" height="10">
+  <svg width="100%" height="10" v-if="orientation === 'horizontal'">
     <line x1="0.5%" y1="50%" x2="99.5%" y2="50%" style="stroke: var(--color-text)" />
     <line
       v-for="([tick, color], i) of ticksAndColors"
@@ -41,6 +45,18 @@ const ticksAndColors = computed(() => {
       y1="5%"
       :x2="tick"
       y2="95%"
+      :style="'stroke:' + color + ';'"
+    />
+  </svg>
+  <svg width="10" height="100%" v-else>
+    <line y1="0.5%" x1="50%" y2="99.5%" x2="50%" style="stroke: var(--color-text)" />
+    <line
+      v-for="([tick, color], i) of ticksAndColors"
+      :key="i"
+      :y1="tick"
+      x1="5%"
+      :y2="tick"
+      x2="95%"
       :style="'stroke:' + color + ';'"
     />
   </svg>